added nsfetch result delegate in the controller
