+++ Artilces/Artilces/ArticleListTableViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import CoreData
 
 class ArticleListTableViewController: UITableViewController, ArticleListDelegate {
 
@@ -26,6 +27,7 @@ class ArticleListTableViewController: UITableViewController, ArticleListDelegate
 		self.tableView.estimatedRowHeight = 250
 		self.tableView.rowHeight = UITableViewAutomaticDimension
 		articleViewModel = ArticleListViewModel.initWith(self)
+		articleViewModel.fetchedhResultController.delegate = self
 		articleViewModel.fetchArticles()
 	}
 
@@ -67,4 +69,25 @@ class ArticleListTableViewController: UITableViewController, ArticleListDelegate
 		UIApplication.shared.open(URL(string : link)!, options: [:], completionHandler: nil)
 	}
 
+}
+
+extension ArticleListTableViewController: NSFetchedResultsControllerDelegate {
+
+	func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
+		switch type {
+		case .insert:
+			self.tableView.insertRows(at: [newIndexPath!], with: .automatic)
+		case .delete:
+			self.tableView.deleteRows(at: [indexPath!], with: .automatic)
+		default:
+			break
+		}
+	}
+	func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
+		self.tableView.endUpdates()
+	}
+
+	func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
+		self.tableView.beginUpdates()
+	}
 }
